console: Move gateway selectors to new file move selectors to selectors/gateways.js
