+++ pkg/webui/console/views/gateway-api-key-edit/index.js
@@ -39,7 +39,7 @@ import {
   selectGatewayApiKey,
   selectGatewayApiKeyError,
   selectGatewayApiKeyFetching,
-} from '../../store/selectors/gateway'
+} from '../../store/selectors/gateways'
 
 import api from '../../api'
