+++ pkg/webui/console/views/gateway-collaborator-edit/index.js
@@ -37,7 +37,7 @@ import {
   selectGatewayUniversalRights,
   selectGatewayRightsFetching,
   selectGatewayRightsError,
-} from '../../store/selectors/gateway'
+} from '../../store/selectors/gateways'
 
 import api from '../../api'
